+++ .eslintrc.json
@@ -12,6 +12,7 @@
     "rules": {
         "prettier/prettier": "error",
         "indent": "off",
-        "consistent-return": "off"
+        "consistent-return": "off",
+        "object-curly-newline": "off"
     }
 }
